Suppression du controlleur Item, déplacement des vues utilisateurs dans User. Suppression css inutiles
