--- public/assets/css/test.css
@@ -1,86 +1,0 @@
-:root {
-    --backColor : lightgrey;
-}
-
-
-body {
-    padding: 4vw;
-    text-align: left;
-
-    background-color: lightgrey;
-    background-color: var(--backColor);
-}
-
-
-header > h1 {
-    margin: 0.3rem;
-    color: darkred;
-}
-
-h2 {
-    margin: 0.3rem;
-    color: darkblue;
-}
-
-body > header { margin-bottom: 3rem }
-
-
-header + hr { border: 1px solid darkred }
-
-
-/* one article per entries in the DB */
-
-
-article {
-    border: 1px solid darkred;
-    border-radius: 0.5rem;
-
-    margin: 2rem;
-    padding: 1rem;
-
-    font-size: 2rem;
-}
-
-
-article > header > h1,
-article > header > h2 {
-    font-size: 1rem;
-}
-
-article > header > h1 {
-    color: darkred;
-    border-bottom: 1px solid darkblue;
-    margin: 0.5rem;
-}
-
-.righted {
-    text-align: right;
-    font-style: italic;
-    color: darkred;
-}
-
-
-
-article > main > p:first-child {
-    color: darkblue;
-    text-align: right;
-    margin: 0.5rem;
-}
-
-
-
-/* error messages */
-
-aside.error { padding: 1rem }
-
-.error {
-    display: inline-block;
-    padding: 0.3rem;
-    margin-top: 1rem;
-
-    color: red;
-    font-weight: bold;
-    border: 1px solid red;
-}
-
-.norm { color: black }
