--- src/Model/Item.php
@@ -1,61 +1,0 @@
-<?php
-/**
- * Created by PhpStorm.
- * User: wcs
- * Date: 23/10/17
- * Time: 10:57
- * PHP version 7
- */
-
-namespace Model;
-
-/**
- * Class Item
- *
- */
-class Item
-{
-    private $id;
-
-    private $title;
-
-    /**
-     * @return int
-     */
-    public function getId(): int
-    {
-        return $this->id;
-    }
-
-    /**
-     * @param mixed $id
-     *
-     * @return Item
-     */
-    public function setId($id): Item
-    {
-        $this->id = $id;
-
-        return $this;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getTitle(): string
-    {
-        return $this->title;
-    }
-
-    /**
-     * @param mixed $title
-     *
-     * @return Item
-     */
-    public function setTitle($title):Item
-    {
-        $this->title = $title;
-
-        return $this;
-    }
-}
