--- src/Model/ItemManager.php
@@ -1,26 +1,0 @@
-<?php
-/**
- * Created by PhpStorm.
- * User: sylvain
- * Date: 07/03/18
- * Time: 18:20
- * PHP version 7
- */
-
-namespace Model;
-
-/**
- *
- */
-class ItemManager extends AbstractManager
-{
-    const TABLE = 'item';
-
-    /**
-     *  Initializes this class.
-     */
-    public function __construct()
-    {
-        parent::__construct(self::TABLE);
-    }
-}
