--- src/View/Item/_form.html.twig
@@ -1,6 +1,0 @@
-<form method="post">
-    <input type="text" name="title">
-    <input type="submit" value="Valider">
-</form>
-
-
