--- src/View/Item/add.html.twig
@@ -1,8 +1,0 @@
-{% extends 'layout.html.twig' %}
-
-{% block content %}
-    <h1>Add a new item</h1>
-    {% include 'Item/_form.html.twig' %}
-{% endblock %}
-
-
