--- src/View/Item/edit.html.twig
@@ -1,8 +1,0 @@
-{% extends 'layout.html.twig' %}
-
-{% block content %}
-    <h1>Edit item {{ item.title }}</h1>
-    {% include 'Item/_form.html.twig' %}
-{% endblock %}
-
-
