--- src/View/Item/show.html.twig
@@ -1,10 +1,0 @@
-{% extends 'layout.html.twig' %}
-
-{% block content %}
-    <h1>Item {{ item.title }}</h1>
-    <p>{{ item.id }}. {{ item.title }}</p>
-    <a href="/item/edit/{{ item.id }}">Edit</a>
-    <a href="/">Back</a>
-{% endblock %}
-
-
